Homework 08 commit
This package contains...
Heap: an abstract class that provided the basic implementation of a heap
and uses an ArrayList to store its data.
MinHeap: a subclass of the abstract class Heap.
MaxHeap: a subclass of the abstract class Heap.
Student: a concrete class that represent a student.
CompareStudentByID extends Heap and implements Comparator that
takes a Student as its parameter.
CompareStudentByName extends Heap and implements Comparator that 
takes a Student as its parameter.
HW08Tester tests Heap, MinHeap, MaxHeap, with instances of the omparator
nterface.