Lab14 and HW09 last commit
This commit contains:
Lab14Programming1 that is a tester for a method find
that find a specified key from an hashTable array.
Lab14 Sefl-Check 2 page 382
Lab14 Sefl-Check 3 page 383
ashtableChain that is a Hash table using chaining 
that fully implements the Map interface.
HashTableChainTester that is a tester class for the 
interface HashTableChain.